backend/requirements.txt
Flask==2.3.3
Flask-RESTX==1.1.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
chromadb==0.4.15
openai==0.28.1
langchain==0.0.329
PyPDF2==3.0.1
pdfplumber==0.9.0
pandas==2.1.3
requests==2.31.0
python-dateutil==2.8.2
uuid==1.30
tenacity==8.2.3
werkzeug==2.3.7
